+++ packages/nesd/test/test_roms/blargg_status_test.dart
@@ -12,6 +12,8 @@ const _passing = <String>[
   'cpu_interrupts_v2/rom_singles/1-cli_latency.nes',
   'mmc3_test/1-clocking.nes',
   'mmc3_test/5-MMC3.nes',
+  'ppu_vbl_nmi/rom_singles/01-vbl_basics.nes',
+  'ppu_vbl_nmi/rom_singles/09-even_odd_frames.nes',
 ];
 
 void main() {